マイグレーションファイル修正のためrollback実行、編集後再度適用しました

+++ db/migrate/20240512145543_drop_recipe_details.rb
@@ -1,6 +1,6 @@
 class DropRecipeDetails < ActiveRecord::Migration[7.1]
   def up
-    drop_table :recipe_details
+    drop_table :recipe_details if ActiveRecord::Base.connection.table_exists? :recipe_details
   end
 
   def down
